+++ mobile/android/android-components/components/feature/readerview/src/test/java/mozilla/components/feature/readerview/ReaderViewFeatureTest.kt
@@ -320,7 +320,7 @@ class ReaderViewFeatureTest {
         testDispatcher.advanceUntilIdle()
 
         readerViewFeature.hideReaderView()
-        verify(engineSession).goBack()
+        verify(engineSession).goBack(false)
     }
 
     @Test